+++ src/components/OSCALProfile.test.js
@@ -2,79 +2,10 @@ import React from "react";
 import { render, screen, fireEvent, act } from "@testing-library/react";
 import { OSCALProfileLoader } from "./OSCALLoader";
 import OSCALProfile from "./OSCALProfile";
-import { metadataTestData, testOSCALMetadata } from "./OSCALMetadata.test";
-import testOSCALBackMatter, {
-  backMatterTestData,
-} from "./OSCALBackMatter.test";
-
-const profileTestData = {
-  uuid: "3afae418-b105-47ba-b51d-653a1a6b9267",
-  metadata: metadataTestData,
-  imports: [
-    {
-      href: "#dc380596-027f-423b-83f2-82757554ee27",
-      "include-controls": [
-        {
-          "with-ids": ["ac-1", "ac-2", "ac-2.1"],
-        },
-      ],
-    },
-  ],
-  modify: {
-    "set-parameters": [
-      {
-        "param-id": "ac-1_prm_2",
-        constraints: [
-          {
-            description: "at least every 3 years",
-          },
-        ],
-      },
-      {
-        "param-id": "ac-1_prm_3",
-        constraints: [
-          {
-            description: "at least annually",
-          },
-        ],
-      },
-    ],
-    alters: [
-      {
-        "control-id": "ac-1",
-        adds: [
-          {
-            position: "starting",
-            props: [
-              {
-                name: "priority",
-                value: "P1",
-              },
-            ],
-          },
-        ],
-      },
-      {
-        "control-id": "ac-2",
-        adds: [
-          {
-            position: "starting",
-            props: [
-              {
-                name: "priority",
-                value: "P1",
-              },
-            ],
-          },
-        ],
-      },
-    ],
-  },
-  "back-matter": backMatterTestData,
-};
-
-const profileParentUrlTestData =
-  "https://raw.githubusercontent.com/usnistgov/oscal-content/master/nist.gov/SP800-53/rev4/json/NIST_SP-800-53_rev4_MODERATE-baseline_profile.json";
+import { testOSCALMetadata } from "./OSCALMetadata.test";
+import testOSCALBackMatter from "./OSCALBackMatter.test";
+import { parentUrlTestData } from "../test-data/Urls";
+import { profileTestData } from "../test-data/OtherData";
 
 test("OSCALProfile loads", () => {
   render(<OSCALProfileLoader />);
@@ -82,10 +13,7 @@ test("OSCALProfile loads", () => {
 
 function profileRenderer() {
   render(
-    <OSCALProfile
-      profile={profileTestData}
-      parentUrl={profileParentUrlTestData}
-    />
+    <OSCALProfile profile={profileTestData} parentUrl={parentUrlTestData} />
   );
 }
