+++ src/components/OSCALSystemCharacteristics.test.js
@@ -1,65 +1,7 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
 import OSCALSystemCharacteristics from "./OSCALSystemCharacteristics";
-
-export const systemCharacteristicsTestData = {
-  "system-name": "Example System Name",
-  description: "This is an example of a system.",
-  "system-ids": [
-    {
-      id: "system-id",
-      "identifier-type": "https://ietf.org/rfc/rfc4122",
-    },
-  ],
-  "security-sensitivity-level": "moderate",
-  "system-information": {
-    "information-types": [
-      {
-        uuid: "information-type-id",
-        title: "Information Type Title",
-        categorizations: [
-          {
-            system: "https://doi.org/10.6028/NIST.SP.800-60v2r1",
-            "information-type-ids": ["C.3.5.8"],
-          },
-        ],
-        description: "Example information type.",
-        "confidentiality-impact": {
-          base: "fips-199-moderate",
-        },
-        "integrity-impact": {
-          base: "fips-199-moderate",
-        },
-        "availability-impact": {
-          base: "fips-199-low",
-        },
-      },
-    ],
-  },
-  "security-impact-level": {
-    "security-objective-confidentiality": "confidentiality-value",
-    "security-objective-integrity": "integrity-value",
-    "security-objective-availability": "availability-value",
-  },
-  status: {
-    state: "other",
-    remarks:
-      "This is an example, and is not intended to be implemented as a system",
-  },
-  props: [
-    {
-      name: "deployment-model",
-      value: "private",
-    },
-    {
-      name: "service-models",
-      value: "iaas",
-    },
-  ],
-  "authorization-boundary": {
-    description: "The description of the authorization boundary would go here.",
-  },
-};
+import { systemCharacteristicsTestData } from "../test-data/SystemData";
 
 function systemCharacteristicsRenderer() {
   render(
