+++ src/stories/OSCALControl.stories.js
@@ -1,5 +1,10 @@
 import React from "react";
 import OSCALControl from "../components/OSCALControl";
+import { exampleControl } from "../test-data/ControlsData";
+import {
+  exampleModificationsAddsThree,
+  exampleModificationsConstraints,
+} from "../test-data/ModificationsData";
 
 export default {
   title: "Components/Control Display",
@@ -12,101 +17,6 @@ export const Default = Template.bind({});
 
 export const AddsModifications = Template.bind({});
 
-const exampleControl = {
-  id: "control-1",
-  title: "Control 1 Title",
-  params: [
-    {
-      id: "control-1_prm_1",
-      label: "control 1 / parameter 1 label",
-    },
-    {
-      id: "control-1_prm_2",
-      label: "control 1 / parameter 2 label",
-    },
-  ],
-  parts: [
-    {
-      id: "control-1_smt",
-      name: "statement",
-      prose: "Some organizational group:",
-      parts: [
-        {
-          id: "control-1_smt.a",
-          name: "item",
-          props: [
-            {
-              name: "label",
-              value: "a.",
-            },
-          ],
-          prose:
-            "Does something with {{ insert: param, control-1_prm_1 }} and {{ insert: param, control-1_prm_2 }}",
-        },
-      ],
-    },
-  ],
-};
-
-const exampleModificationsConstraints = {
-  "set-parameters": [
-    {
-      "param-id": "control-1_prm_1",
-      constraints: [
-        {
-          description: "some constraint",
-        },
-      ],
-    },
-    {
-      "param-id": "control-1_prm_2",
-      constraints: [
-        {
-          description: "another constraint",
-        },
-      ],
-    },
-  ],
-};
-
-const exampleModificationsAdds = {
-  alters: [
-    {
-      "control-id": "control-1",
-      adds: [
-        {
-          position: "starting",
-          props: [
-            {
-              name: "Name",
-              value: "Value",
-            },
-          ],
-        },
-      ],
-    },
-  ],
-
-  "set-parameters": [
-    {
-      "param-id": "control-1_prm_1",
-      constraints: [
-        {
-          description: "some constraint",
-        },
-      ],
-    },
-    {
-      "param-id": "control-1_prm_2",
-      constraints: [
-        {
-          description: "another constraint",
-        },
-      ],
-    },
-  ],
-};
-
 Default.args = {
   control: exampleControl,
   modifications: exampleModificationsConstraints,
@@ -114,5 +24,5 @@ Default.args = {
 
 AddsModifications.args = {
   control: exampleControl,
-  modifications: exampleModificationsAdds,
+  modifications: exampleModificationsAddsThree,
 };
