+++ src/test-data/Urls.js
@@ -1,0 +1,13 @@
+/* The following URLs are all of the URLs used in testing and in the stories. */
+export const revFourCatalog =
+  "https://raw.githubusercontent.com/usnistgov/oscal-content/master/nist.gov/SP800-53/rev4/json/NIST_SP-800-53_rev4_catalog.json";
+export const parentUrlTestData =
+  "https://raw.githubusercontent.com/usnistgov/oscal-content/master/nist.gov/SP800-53/rev4/json/NIST_SP-800-53_rev4_MODERATE-baseline_profile.json";
+export const backMatterTestUrl =
+  "https://raw.githubusercontent.com/usnistgov/oscal-content/master/nist.gov/SP800-53/rev4/json/NIST_SP-800-53_rev4_MODERATE-baseline_profile.json/../NIST_SP-800-53_rev4_catalog.json";
+export const fedRampHighBaselineUrl =
+  "https://raw.githubusercontent.com/usnistgov/oscal-content/master/fedramp.gov/json/FedRAMP_HIGH-baseline_profile.json";
+export const systemCharacteristicsDescriptionUrl =
+  "https://ietf.org/rfc/rfc4122";
+export const systemCharacteristicsInformationUrl =
+  "https://doi.org/10.6028/NIST.SP.800-60v2r1";
